Moved VertexFormat
