+++ naga/src/back/msl/mod.rs
@@ -338,105 +338,6 @@ impl Default for Options {
     }
 }
 
-/// Corresponds to [WebGPU `GPUVertexFormat`](
-/// https://gpuweb.github.io/gpuweb/#enumdef-gpuvertexformat).
-#[repr(u32)]
-#[derive(Copy, Clone, Debug, Hash, Eq, PartialEq)]
-#[cfg_attr(feature = "serialize", derive(serde::Serialize))]
-#[cfg_attr(feature = "deserialize", derive(serde::Deserialize))]
-pub enum VertexFormat {
-    /// One unsigned byte (u8). `u32` in shaders.
-    Uint8 = 0,
-    /// Two unsigned bytes (u8). `vec2<u32>` in shaders.
-    Uint8x2 = 1,
-    /// Four unsigned bytes (u8). `vec4<u32>` in shaders.
-    Uint8x4 = 2,
-    /// One signed byte (i8). `i32` in shaders.
-    Sint8 = 3,
-    /// Two signed bytes (i8). `vec2<i32>` in shaders.
-    Sint8x2 = 4,
-    /// Four signed bytes (i8). `vec4<i32>` in shaders.
-    Sint8x4 = 5,
-    /// One unsigned byte (u8). [0, 255] converted to float [0, 1] `f32` in shaders.
-    Unorm8 = 6,
-    /// Two unsigned bytes (u8). [0, 255] converted to float [0, 1] `vec2<f32>` in shaders.
-    Unorm8x2 = 7,
-    /// Four unsigned bytes (u8). [0, 255] converted to float [0, 1] `vec4<f32>` in shaders.
-    Unorm8x4 = 8,
-    /// One signed byte (i8). [-127, 127] converted to float [-1, 1] `f32` in shaders.
-    Snorm8 = 9,
-    /// Two signed bytes (i8). [-127, 127] converted to float [-1, 1] `vec2<f32>` in shaders.
-    Snorm8x2 = 10,
-    /// Four signed bytes (i8). [-127, 127] converted to float [-1, 1] `vec4<f32>` in shaders.
-    Snorm8x4 = 11,
-    /// One unsigned short (u16). `u32` in shaders.
-    Uint16 = 12,
-    /// Two unsigned shorts (u16). `vec2<u32>` in shaders.
-    Uint16x2 = 13,
-    /// Four unsigned shorts (u16). `vec4<u32>` in shaders.
-    Uint16x4 = 14,
-    /// One signed short (u16). `i32` in shaders.
-    Sint16 = 15,
-    /// Two signed shorts (i16). `vec2<i32>` in shaders.
-    Sint16x2 = 16,
-    /// Four signed shorts (i16). `vec4<i32>` in shaders.
-    Sint16x4 = 17,
-    /// One unsigned short (u16). [0, 65535] converted to float [0, 1] `f32` in shaders.
-    Unorm16 = 18,
-    /// Two unsigned shorts (u16). [0, 65535] converted to float [0, 1] `vec2<f32>` in shaders.
-    Unorm16x2 = 19,
-    /// Four unsigned shorts (u16). [0, 65535] converted to float [0, 1] `vec4<f32>` in shaders.
-    Unorm16x4 = 20,
-    /// One signed short (i16). [-32767, 32767] converted to float [-1, 1] `f32` in shaders.
-    Snorm16 = 21,
-    /// Two signed shorts (i16). [-32767, 32767] converted to float [-1, 1] `vec2<f32>` in shaders.
-    Snorm16x2 = 22,
-    /// Four signed shorts (i16). [-32767, 32767] converted to float [-1, 1] `vec4<f32>` in shaders.
-    Snorm16x4 = 23,
-    /// One half-precision float (no Rust equiv). `f32` in shaders.
-    Float16 = 24,
-    /// Two half-precision floats (no Rust equiv). `vec2<f32>` in shaders.
-    Float16x2 = 25,
-    /// Four half-precision floats (no Rust equiv). `vec4<f32>` in shaders.
-    Float16x4 = 26,
-    /// One single-precision float (f32). `f32` in shaders.
-    Float32 = 27,
-    /// Two single-precision floats (f32). `vec2<f32>` in shaders.
-    Float32x2 = 28,
-    /// Three single-precision floats (f32). `vec3<f32>` in shaders.
-    Float32x3 = 29,
-    /// Four single-precision floats (f32). `vec4<f32>` in shaders.
-    Float32x4 = 30,
-    /// One unsigned int (u32). `u32` in shaders.
-    Uint32 = 31,
-    /// Two unsigned ints (u32). `vec2<u32>` in shaders.
-    Uint32x2 = 32,
-    /// Three unsigned ints (u32). `vec3<u32>` in shaders.
-    Uint32x3 = 33,
-    /// Four unsigned ints (u32). `vec4<u32>` in shaders.
-    Uint32x4 = 34,
-    /// One signed int (i32). `i32` in shaders.
-    Sint32 = 35,
-    /// Two signed ints (i32). `vec2<i32>` in shaders.
-    Sint32x2 = 36,
-    /// Three signed ints (i32). `vec3<i32>` in shaders.
-    Sint32x3 = 37,
-    /// Four signed ints (i32). `vec4<i32>` in shaders.
-    Sint32x4 = 38,
-    /// Three unsigned 10-bit integers and one 2-bit integer, packed into a 32-bit integer (u32). [0, 1024] converted to float [0, 1] `vec4<f32>` in shaders.
-    #[cfg_attr(
-        any(feature = "serialize", feature = "deserialize"),
-        serde(rename = "unorm10-10-10-2")
-    )]
-    Unorm10_10_10_2 = 43,
-    /// Four unsigned 8-bit integers, packed into a 32-bit integer (u32). [0, 255] converted to float [0, 1] `vec4<f32>` in shaders.
-    #[cfg_attr(
-        any(feature = "serialize", feature = "deserialize"),
-        serde(rename = "unorm8x4-bgra")
-    )]
-    Unorm8x4Bgra = 44,
-}
-
 /// Defines how to advance the data in vertex buffers.
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Hash)]
 #[cfg_attr(feature = "serialize", derive(serde::Serialize))]
